requirements.txt
web3>=6.0.0
python-dotenv>=0.19.0
rich>=12.0.0
click>=8.0.0
